fixed bugs - app crashed when value wasn't supplied to number of employees text field

+++ app/src/main/java/com/decagonhq/clads/ui/view/profilemanagementfragments/dialogfragments/EditProfileAccountNumberOfEmployeesCustomDialogFragment.kt
@@ -33,7 +33,7 @@ class EditProfileAccountNumberOfEmployeesCustomDialogFragment : DialogFragment()
 
         binding.fragmentProfileAccountNumberOfEmployeesCustomDialogOkTextView.setOnClickListener {
             val numberOfEmployees = binding.fragmentProfileAccountNumberOfEmployeesCustomDialogNumberOfEmployeesEditText.editText?.text.toString()
-            viewModel.numberOfEmployees.value = numberOfEmployees.toInt()
+            viewModel.numberOfEmployees.value = numberOfEmployees
             dismiss()
         }
